rats

+++ Source/RatageddonAR/Assets/Enemy/Scripts/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Enemy.Scripts
 {
     public interface IDamageable
     {
-        public void GetDamage(float damage);
+        public void GetDamage(int damage);
     }
 }
